feat: [ANDROAPP-5125] Refactor RelationshipsRepository from an interface to an abstract class
